proposal: workspace/documentation support
The aim of this extension to the LSP specification is to enable language servers
to produce API documentation that is competitive with the API documentation offered
in some languages today through various websites such as:

* (Go) https://pkg.go.dev
* (Rust) https://docs.rs
* (Java) https://javadoc.io
* (Zig) https://ziglang.org/documentation/master/std/#builtin

By enabling language servers to emit structured documentation, and link specific
ranges in that documentation to other documentation, hover-able ranges, go-to-definition
requests, and find-references requests it will be possible to decouple the _backend
implementation_ of API docs for such a website from its frontend implementation,
hopefully also making it far easier for languages to merely export the relevant
documentation data they care about, and defer providing a nice user experience for viewing
and navigating that documentation to others (such as Sourcegraph.)

Of course, this being an extension to LSP means it will primarily only be suitable
in editor environments. This would enable developers to get nice API documentation
similar to the previously mentioned websites directly in their editors. For an actual
website based on this data, storing it in the LSIF format would likely be a requirement.

Helps https://github.com/sourcegraph/sourcegraph/issues/19389

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>